requirements.txt
numpy
matplotlib
scipy
tqdm
threadpoolctl
scikit-learn
umap-learn
openai
python-dotenv
torch